rename plugin into flytekit-omegaconf
