+++ plugins/flytekit-omegaconf/flytekitplugins/omegaconf/__init__.py
@@ -1,0 +1,3 @@
+from flytekitplugins.omegaconf.dictconfig_transformer import DictConfigTransformer  # noqa: F401
+from flytekitplugins.omegaconf.extended_enum_transformer import GenericEnumTransformer  # noqa: F401
+from flytekitplugins.omegaconf.listconfig_transformer import ListConfigTransformer  # noqa: F401
