+++ plugins/flytekit-omegaconf/flytekitplugins/omegaconf/listconfig_transformer.py
@@ -5,9 +5,9 @@ from typing import Optional, Type, TypeVar
 
 import omegaconf
 from flyteidl.core.literals_pb2 import Literal as PB_Literal
-from flytekitplugins.hydra.config import OmegaConfTransformerMode, SharedConfig
-from flytekitplugins.hydra.flytekit_patch import iterate_get_transformers
-from flytekitplugins.hydra.type_information import extract_node_type
+from flytekitplugins.omegaconf.config import OmegaConfTransformerMode, SharedConfig
+from flytekitplugins.omegaconf.flytekit_patch import iterate_get_transformers
+from flytekitplugins.omegaconf.type_information import extract_node_type
 from google.protobuf.json_format import MessageToDict, ParseDict
 from google.protobuf.struct_pb2 import Struct
 from omegaconf import ListConfig, OmegaConf
